+++ src/OpenFPL_frontend/src/lib/managers/store-manager.ts
@@ -4,7 +4,6 @@ import { clubStore } from "$lib/stores/club-store";
 import { playerStore } from "$lib/stores/player-store";
 import { playerEventsStore } from "$lib/stores/player-events-store";
 import { fixtureStore } from "$lib/stores/fixture-store";
-import { weeklyLeaderboardStore } from "$lib/stores/leaderboard-store";
 import { leagueStore } from "$lib/stores/league-store";
 import { appStore } from "$lib/stores/app-store";
 import { RewardRatesService } from "$lib/services/reward-rates-service";
@@ -19,8 +18,6 @@ import { ClubService } from "$lib/services/club-service";
 import { PlayerService } from "$lib/services/player-service";
 import { PlayerEventsService } from "$lib/services/player-events-service";
 import { FixtureService } from "$lib/services/fixture-service";
-import { WeeklyLeaderboardService } from "$lib/services/leaderboard-service";
-
 import { replacer } from "$lib/utils/helpers";
 import { writable } from "svelte/store";
 
@@ -36,7 +33,6 @@ class StoreManager {
   private playerService: PlayerService;
   private playerEventsService: PlayerEventsService;
   private fixtureService: FixtureService;
-  private weeklyLeaderboardService: WeeklyLeaderboardService;
   private rewardRatesService: RewardRatesService;
 
   private backendCategories: string[] = [
@@ -64,7 +60,6 @@ class StoreManager {
     this.playerService = new PlayerService();
     this.playerEventsService = new PlayerEventsService();
     this.fixtureService = new FixtureService();
-    this.weeklyLeaderboardService = new WeeklyLeaderboardService();
     this.rewardRatesService = new RewardRatesService();
   }
 
@@ -206,30 +201,6 @@ class StoreManager {
           JSON.stringify(updatedFixtures.fixtures, replacer),
         );
         break;
-      case "weekly_leaderboard":
-        leagueStore.subscribe(async (leagueStatus) => {
-          if (!leagueStatus) {
-            return;
-          }
-          const updatedWeeklyLeaderboard =
-            await this.weeklyLeaderboardService.getWeeklyLeaderboard(
-              0,
-              leagueStatus.activeSeasonId,
-              leagueStatus.activeGameweek == 0
-                ? leagueStatus.completedGameweek
-                : leagueStatus.activeGameweek,
-              "",
-            );
-          if (!updatedWeeklyLeaderboard) {
-            return;
-          }
-          weeklyLeaderboardStore.setWeeklyLeaderboard(updatedWeeklyLeaderboard);
-          localStorage.setItem(
-            "weekly_leaderboard",
-            JSON.stringify(updatedWeeklyLeaderboard, replacer),
-          );
-        });
-        break;
       case "reward_rates":
         leagueStore.subscribe(async (leagueStatus) => {
           if (!leagueStatus) {
@@ -281,10 +252,6 @@ class StoreManager {
       case "fixtures":
         const cachedFixtures = JSON.parse(cachedData || "[]");
         fixtureStore.setFixtures(cachedFixtures);
-        break;
-      case "weekly_leaderboard":
-        const cachedWeeklyLeaderboard = JSON.parse(cachedData || "null");
-        weeklyLeaderboardStore.setWeeklyLeaderboard(cachedWeeklyLeaderboard);
         break;
       case "reward_rates":
         const cachedRewardRates = JSON.parse(cachedData || "null");
