+++ src/OpenFPL_frontend/src/routes/(private)/+page.svelte
@@ -1,24 +1,16 @@
 <script lang="ts">
-  import { onMount } from "svelte";
-  
-  import { seasonStore } from "$lib/stores/season-store";
-  import { leagueStore } from "$lib/stores/league-store";
-  import { globalDataLoaded } from "$lib/managers/store-manager";
-  import { appStore } from "$lib/stores/app-store";
-  import { userStore } from "$lib/stores/user-store";
-  import { storeManager } from "$lib/managers/store-manager";
   
   import HomepageHeader from "$lib/components/homepage/homepage-header.svelte";
-  import FixturesComponent from "$lib/components/homepage/fixtures.svelte";
-  import GamweekPointsComponent from "$lib/components/manager/gameweek-points.svelte";
-  import LeaderboardsComponent from "$lib/components/homepage/leaderboards.svelte";
-  import LeagueTableComponent from "$lib/components/homepage/league-table.svelte";
-  import TabContainer from "$lib/components/shared/tab-container.svelte";
-  import LocalSpinner from "$lib/components/shared/local-spinner.svelte";
 
-  let activeTab: string = "fixtures";
-  let isLoading = true;
-  let seasonName = "";
+  import TabContainer from "$lib/components/shared/tab-container.svelte";
+  import MyPoints from "$lib/components/home/my-points.svelte";
+  import LeaderboardsComponent from "$lib/components/home/leaderboards.svelte";
+  import FixturesComponent from "$lib/components/home/fixtures.svelte";
+  import MvpsComponent from "$lib/components/home/mvps.svelte";
+  import LeagueTableComponent from "$lib/components/home/league-table.svelte";
+  import TrophyRoomComponent from "$lib/components/home/trophy-room.svelte";
+
+  let activeTab: string = "points";
 
   const tabs = [
     { id: "points", label: "My Points" },
@@ -29,48 +21,28 @@
     { id: "trophy-room", label: "Trophy Room" },
   ];
 
-  onMount(() => {
-    let unsub: () => void = () => {};
-    unsub = globalDataLoaded.subscribe((loaded) => {
-      if (loaded) {
-        loadCurrentStatusDetails();
-        isLoading = false;
-        unsub();
-      }
-    });
-    isLoading = false;
-  });
-
-  async function loadCurrentStatusDetails(){
-    seasonName = await seasonStore.getSeasonName($leagueStore?.activeSeasonId ?? 1) ?? "-";
-  }
-
   function setActiveTab(tab: string): void {
     activeTab = tab;
   }
 
 </script>
 
-{#if isLoading}
-    <LocalSpinner />
-{:else}
-    <HomepageHeader {seasonName} />
+<HomepageHeader />
 
-    <div class="bg-panel">
-      <TabContainer {tabs} {activeTab} {setActiveTab} />
-      
-      {#if activeTab === "points"}
-        <GamweekPointsComponent />
-      {:else if activeTab === "leaderboards"}
-        <LeaderboardsComponent />
-      {:else if activeTab === "fixtures"}
-      <FixturesComponent />
-      {:else if activeTab === "mvps"}
-        <MVPsComponent />
-      {:else if activeTab === "league-table"}
-        <LeagueTableComponent />
-      {:else if activeTab === "trophy-room"}
-        <TrophyRoomComponent />
-      {/if}
-    </div>
-{/if}
+<div class="bg-panel">
+  <TabContainer {tabs} {activeTab} {setActiveTab} />
+  
+  {#if activeTab === "points"}
+    <MyPoints />
+  {:else if activeTab === "leaderboards"}
+    <LeaderboardsComponent />
+  {:else if activeTab === "fixtures"}
+  <FixturesComponent />
+  {:else if activeTab === "mvps"}
+    <MvpsComponent />
+  {:else if activeTab === "league-table"}
+    <LeagueTableComponent />
+  {:else if activeTab === "trophy-room"}
+    <TrophyRoomComponent />
+  {/if}
+</div>
